revoice-back/requirements.txt
fastapi
uvicorn
aiofiles
python-dotenv
requests
faster-whisper
ffmpeg-python
python-multipart
